:bug: Update base api basepath
This, in combination with the Apache WSGIScriptAlias, was causing the
web app to be incorrectly mounted.

Also allowed the API to differentiate between queries from the archive
viewer portion of the AVERT data portal, and requests made directly to
the API from a tool such as cURL or the Python requests package.